Improve styling for AccountInformation page
