+++ src/components/List/styles.module.css
@@ -1,0 +1,9 @@
+.message {
+    &.with-message-over-content {
+        position: absolute;
+        top: 0;
+        left: 0;
+        width: 100%;
+        height: 100%;
+    }
+}
