--- src/views/Account/AccountInfo/OperationInfoCard/styles.module.css
@@ -1,19 +1,0 @@
-.operations-box {
-    display: flex;
-    flex-direction: column;
-    border-radius: var(--go-ui-border-radius-md);
-    box-shadow: var(--go-ui-box-shadow-lg);
-    padding: var(--go-ui-spacing-sm);
-
-    .header {
-        padding: var(--go-ui-spacing-md);
-    }
-
-    .last-updated {
-        display: flex;
-        padding: 0 var(--go-ui-spacing-md);
-        padding: var(--go-ui-spacing-md);
-        color: var(--go-ui-color-gray-60);
-        gap: var(--go-ui-spacing-sm);
-    }
-}
