--- src/views/Account/AccountInfo/styles.module.css
@@ -1,50 +1,0 @@
-.account-info {
-    /* display: flex;
-    padding: var(--go-ui-spacing-2xl) 0;
-    gap: var(--go-ui-spacing-2xl); */
-    display: grid;
-    grid-gap: var(--go-ui-spacing-2xl);
-    grid-template-columns: repeat(auto-fit, minmax(16rem, 1fr));
-    padding: var(--go-ui-spacing-2xl) 0;
-
-    .info-box {
-        display: flex;
-        flex-direction: column;
-        border-radius: var(--go-ui-border-radius-md);
-        box-shadow: var(--go-ui-box-shadow-lg);
-        padding: var(--go-ui-spacing-2xl);
-        width: 100%;
-        max-height: 40rem;
-
-        .operations-list {
-            display: grid;
-            grid-gap: var(--go-ui-spacing-4xl);
-            grid-template-columns: repeat(auto-fit, minmax(26rem, 1fr));
-        }
-
-        .user-detail-section {
-            display: flex;
-            flex-direction: column;
-            font-size: var(--go-ui-font-size-lg);
-
-            .user-info-row {
-                padding: var(--go-ui-spacing-xs);
-            }
-
-            .user-info {
-                font-weight: var(--go-ui-font-weight-semibold);
-            }
-        }
-
-        .change-password-button {
-            padding: var(--go-ui-spacing-md) 0;
-            text-decoration: underline;
-        }
-    }
-
-    .user-edit {
-        display: flex;
-        flex-wrap: wrap;
-        gap: var(--go-ui-spacing-sm);
-    }
-}
