+++ src/views/AccountInformation/ChangePassword/styles.module.css
@@ -1,0 +1,7 @@
+.change-password {
+    .content {
+        display: flex;
+        flex-direction: column;
+        gap: var(--go-ui-spacing-xl);
+    }
+}
