+++ src/views/AccountInformation/EditAccountInfo/styles.module.css
@@ -1,0 +1,17 @@
+.edit-account-info {
+    .content {
+        display: grid;
+        grid-gap: var(--go-ui-spacing-2xl);
+        grid-template-columns: repeat(auto-fit, minmax(12rem, 1fr));
+        background-color: var(--go-ui-color-white);
+
+        .email-input {
+            grid-column: 1 / -1;
+
+        }
+
+        .non-field-error {
+            grid-column: 1 / -1;
+        }
+    }
+}
