+++ src/views/AccountInformation/OperationInfoCard/styles.module.css
@@ -1,0 +1,3 @@
+.operations-info-card {
+    color: inherit;
+}
